Validate pbkdf2 settings are compliant in FIPS mode
